requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
openai==1.3.7
httpx==0.25.2
python-dotenv==1.0.0
sqlalchemy==2.0.23
pymysql==1.1.0
sshtunnel==0.4.0